requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.9.0
pydantic-settings>=2.5.0
torch>=2.6.0
torchvision>=0.21.0
ultralytics==8.1.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy>=1.26.0
Pillow>=10.0.0
scikit-learn>=1.4.0
imageio>=2.33.0
imageio-ffmpeg==0.4.9
geopy==2.4.1
folium==0.15.1
geopandas==0.14.1
shapely>=2.0.3
pandas>=2.2.0
scipy>=1.12.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
reportlab==4.0.7
fpdf2==2.7.6
jinja2==3.1.2
sqlalchemy==2.0.23
alembic==1.12.1
supabase>=2.8.0
postgrest>=0.14.0
websockets>=12.0
# Pin httpx to satisfy gotrue and supafunc constraints
httpx==0.25.2
python-dotenv>=1.0.0
pyyaml==6.0.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1

# Development / tooling
# Black - code formatter
black==23.12.0